Replacement of 3D Overlays for Local Entities
This replaces all the deprecated 3d Overlays for Local Entities.
No functional changes, only technical.

Many variables have been renamed for clarity. 
(Mainly the term "overlay" that were used to designate the elements of the selection tool. 
Now it refer to "toolEntity" to designate the entities composing the Selection Tool.)

The color of the different elements of the selection tools has also been adjusted.
They are now more contrasted and more vivid.
(The green and the blue of the previous version were too much close at the point that 
some color blind people might have difficulty to differenciate them.)